Compilation avec numéro de joueurs / gestion de ou récupérer les coups

+++ chessminitel/include/tools/LoRa.h
@@ -12,14 +12,14 @@ extern "C" {
 #include <string.h>
 #include <inttypes.h>
 
-#include "../include/utils_LoRa.h"
+#include "utils_LoRa.h"
 
 #define MESSAGE_LENGTH 4
 #define LORA_MESSAGE_BUFFER_SIZE 64
 
 void initialize_lora(void);
-void listen_for_message(uint8_t* coup_recu);
-void send_lora_message(uint8_t* coup_a_envoyer);
+void listen_for_message(int* coup_recu);
+void send_lora_message(int* coup_a_envoyer);
 
 #ifdef __cplusplus
 }
